# Conflicts:
#	DefaultWindow/Bullet.cpp
#	DefaultWindow/ChargeBullet.cpp
#	DefaultWindow/ChargeBullet.h
#	DefaultWindow/GameMgr.cpp
#	DefaultWindow/GuideBullet.cpp
#	DefaultWindow/GuideBullet.h
#	DefaultWindow/Line.cpp
#	DefaultWindow/Line.h
#	DefaultWindow/LineMgr.cpp
#	DefaultWindow/LineMgr.h
#	DefaultWindow/MainGame.cpp
#	DefaultWindow/Monster_Stage1.cpp
#	DefaultWindow/Monster_Stage1B.cpp
#	DefaultWindow/Monster_Stage1Boss.cpp
#	DefaultWindow/Monster_Stage1Boss.h
#	DefaultWindow/ObjMgr.cpp
#	DefaultWindow/ObjMgr.h
#	DefaultWindow/Player.cpp
#	DefaultWindow/Player_Stage3.cpp
#	DefaultWindow/Player_Stage3.h
#	DefaultWindow/ShotgunBullet.cpp
#	DefaultWindow/ShotgunBullet.h
#	DefaultWindow/Spawn.cpp
#	DefaultWindow/Spawn.h
#	DefaultWindow/Stage1.cpp
#	DefaultWindow/Stage2.cpp
#	Header/Define.h

+++ DefaultWindow/Player.cpp
@@ -46,7 +46,7 @@ void CPlayer::Late_Update()
 
 void CPlayer::Render(HDC hDC)
 {
-	int iScrollX = CScrollMgr::Get_Instance()->Get_ScrollX();
+	int iScrollX = (int)CScrollMgr::Get_Instance()->Get_ScrollX();
 	Rectangle(hDC,
 		m_tRect.left + iScrollX,
 		m_tRect.top,
@@ -113,8 +113,8 @@ void CPlayer::HandleMove()
 
 void CPlayer::OffSet()
 {
-	int	iOffSetMinX = 100.f;
-	int	iOffSetMaxX = 700.f;
+	int	iOffSetMinX = 100;
+	int	iOffSetMaxX = 700;
 
 	int iScrollX = (int)CScrollMgr::Get_Instance()->Get_ScrollX();
